fix clang warning

+++ menu.c
@@ -919,9 +919,9 @@ void TeletextBrowser::UpdateFooter() {
       DEBUG_OT_FOOT("AkRed=%d AkGreen=%d AkYellow=%d AkBlue=%d", AkRed, AkGreen, AkYellow, AkBlue);
 
 #define CONVERT_ACTION_TO_TEXT(text, mode) \
-      if (mode < 100) { \
+      if ((int) mode < 100) { \
          snprintf(text, sizeof(text), "%s", tr(st_modesFooter[mode])); \
-      } else if (mode < 999) { \
+      } else if ((int) mode < 999) { \
          snprintf(text, sizeof(text), "-> %03d", mode); \
       } else { \
          snprintf(text, sizeof(text), "ERROR"); \
